Add JWT authentication with protected routes

+++ main.py
@@ -2,9 +2,10 @@ from fastapi import FastAPI, Depends, HTTPException
 from sqlalchemy.orm import Session
 from pydantic import BaseModel
 from typing import Optional, List
-
+from auth import verify_password, create_access_token
 import models
 from database import engine, SessionLocal
+from auth import get_current_user
 
 # Create database tables
 models.Base.metadata.create_all(bind=engine)
@@ -12,6 +13,7 @@ models.Base.metadata.create_all(bind=engine)
 app = FastAPI()
 
 # DB session dependency
+
 def get_db():
     db = SessionLocal()
     try:
@@ -20,6 +22,7 @@ def get_db():
         db.close()
 
 # ----- Schemas -----
+
 class TaskBase(BaseModel):
     title: str
     description: Optional[str] = None
@@ -35,6 +38,7 @@ class Task(TaskBase):
         from_attributes = True
 
 # new: TaskUpdate schema for partial updates
+
 class TaskUpdate(BaseModel):
     title: Optional[str] = None
     description: Optional[str] = None
@@ -57,20 +61,45 @@ class UserUpdate(BaseModel):
     username: Optional[str] = None
     password: Optional[str] = None
 
+#Schema for auth
 
+class LoginInput(BaseModel):
+    username: str
+    password: str
+
+#---------LOGIN Enpoint---------
+@app.post("/login")
+def login(user_input: LoginInput, db: Session = Depends(get_db)):
+    user = db.query(models.User).filter(models.User.username == user_input.username).first()
+
+    if not user:
+        raise HTTPException(status_code=400, detail="Invalid username or password")
+
+    if not verify_password(user_input.password, user.password):
+        raise HTTPException(status_code=400, detail="Invalid username or password")
+
+    # create token
+    token = create_access_token({"sub": str(user.id)})
+
+    return {"access_token": token, "token_type": "bearer"}
 
 # ----- User Endpoints -----
+from auth import hash_password
+
 @app.post("/users", response_model=UserOut)
 def create_user(user: UserCreate, db: Session = Depends(get_db)):
     db_user = db.query(models.User).filter(models.User.username == user.username).first()
     if db_user:
-        raise HTTPException(status_code=400, detail="Username already registered")
+        raise HTTPException(status_code=400, detail="Username already exists")
 
-    new_user = models.User(username=user.username, password=user.password)
+    hashed_pass = hash_password(user.password)
+
+    new_user = models.User(username=user.username, password=hashed_pass)
     db.add(new_user)
     db.commit()
     db.refresh(new_user)
     return new_user
+
 
 
 @app.get("/users/{user_id}", response_model=UserOut)
@@ -109,17 +138,24 @@ def delete_user(user_id: int, db: Session = Depends(get_db)):
     return {"message": f"User {user_id} deleted successfully"}
 
 # ----- Task Endpoints -----
-@app.post("/users/{user_id}/tasks", response_model=Task)
-def create_task_for_user(user_id: int, task: TaskCreate, db: Session = Depends(get_db)):
-    db_user = db.query(models.User).filter(models.User.id == user_id).first()
-    if not db_user:
-        raise HTTPException(status_code=404, detail="User not found")
 
-    db_task = models.Task(**task.dict(), owner_id=user_id)
-    db.add(db_task)
+@app.post("/tasks", response_model=Task)
+def create_task(
+    task: TaskCreate,
+    db: Session = Depends(get_db),
+    current_user: models.User = Depends(get_current_user)
+):
+    new_task = models.Task(
+        title=task.title,
+        description=task.description,
+        completed=task.completed,
+        owner_id=current_user.id
+    )
+
+    db.add(new_task)
     db.commit()
-    db.refresh(db_task)
-    return db_task
+    db.refresh(new_task)
+    return new_task
 
 @app.get("/tasks", response_model=List[Task])
 def read_tasks(db: Session = Depends(get_db)):
@@ -134,11 +170,20 @@ def read_task(task_id: int, db: Session = Depends(get_db)):
 
 #PUT is both PUT and PATCH here
 @app.put("/tasks/{task_id}", response_model=Task)
-def update_task(task_id: int, updated_task: TaskCreate, db: Session = Depends(get_db)):
+def update_task(
+    task_id: int,
+    updated_task: TaskUpdate,
+    db: Session = Depends(get_db),
+    current_user: models.User = Depends(get_current_user)
+):
     task = db.query(models.Task).filter(models.Task.id == task_id).first()
+
     if not task:
         raise HTTPException(status_code=404, detail="Task not found")
-   
+
+    if task.owner_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Not authorized")
+
     if updated_task.title is not None:
         task.title = updated_task.title
     if updated_task.description is not None:
@@ -151,11 +196,19 @@ def update_task(task_id: int, updated_task: TaskCreate, db: Session = Depends(ge
     return task
 
 @app.delete("/tasks/{task_id}")
-def delete_task(task_id: int, db: Session = Depends(get_db)):
+def delete_task(
+    task_id: int,
+    db: Session = Depends(get_db),
+    current_user: models.User = Depends(get_current_user)
+):
     task = db.query(models.Task).filter(models.Task.id == task_id).first()
+
     if not task:
         raise HTTPException(status_code=404, detail="Task not found")
 
+    if task.owner_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Not authorized")
+
     db.delete(task)
     db.commit()
-    return {"message": "Task deleted successfully"}
+    return {"message": "Task deleted"}
